Start moving functions to standalone folders
