+++ src/lib/get-configuration/index.js
@@ -1,24 +1,5 @@
 const chalk = require('chalk');
-const { join } = require('path');
-
-const getServePath = (dir, subDir) => {
-  if (typeof subDir !== 'string' || typeof dir !== 'string') {
-    return { serveDir: undefined };
-  }
-
-  const resolvedPath = join(dir, subDir);
-  if (!resolvedPath.startsWith(dir)) {
-    throw new Error(
-      chalk.red(
-        `resolved path for ${chalk.red(
-          subDir,
-        )} is outside publish directory ${chalk.red(dir)}`,
-      ),
-    );
-  }
-
-  return { serveDir: resolvedPath };
-};
+const getServePath = require('../get-serve-path');
 
 const getConfiguration = ({ constants, inputs } = {}) => {
   const serveDir =
@@ -72,6 +53,4 @@ const getConfiguration = ({ constants, inputs } = {}) => {
   return { audits };
 };
 
-module.exports = {
-  getConfiguration,
-};
+module.exports = getConfiguration;
